Added initial happy path logic, reading instructions after branching, and minor stuff
Added the initial happy path logic to move instructions from one stage
to another. Should be in a format that makes it easy to squash and
stall. Also put in the beginnings of a queue to handle instructions that
we read from the trace, but were backed out because of a branch. Also
had a few minor things.